[AIRFLOW-5196] Move Google DLP to core
This commit moves GCP DLP from contrib to core.
For more information check AIP-21.
