+++ tests/gcp/hooks/test_dlp.py
@@ -28,7 +28,7 @@ from typing import Any, Dict
 from google.cloud.dlp_v2.types import DlpJob
 
 from airflow import AirflowException
-from airflow.contrib.hooks.gcp_dlp_hook import CloudDLPHook
+from airflow.gcp.hooks.dlp import CloudDLPHook
 from tests.compat import mock
 from tests.contrib.utils.base_gcp_mock import mock_base_gcp_hook_no_default_project_id
 
@@ -66,13 +66,13 @@ JOB_TRIGGER_PATH = "projects/{}/jobTriggers/{}".format(PROJECT_ID, TRIGGER_ID)
 class TestCloudDLPHook(unittest.TestCase):
     def setUp(self):
         with mock.patch(
-            "airflow.contrib.hooks." "gcp_api_base_hook.GoogleCloudBaseHook.__init__",
+            "airflow.contrib.hooks.gcp_api_base_hook.GoogleCloudBaseHook.__init__",
             new=mock_base_gcp_hook_no_default_project_id,
         ):
             self.hook = CloudDLPHook(gcp_conn_id="test")
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_cancel_dlp_job(self, get_conn):
         self.hook.cancel_dlp_job(dlp_job_id=DLP_JOB_ID, project_id=PROJECT_ID)
@@ -82,21 +82,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_cancel_dlp_job_without_dlp_job_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.cancel_dlp_job(dlp_job_id=None, project_id=PROJECT_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_cancel_dlp_job_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.cancel_dlp_job(dlp_job_id=DLP_JOB_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.create_deidentify_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -115,7 +115,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.create_deidentify_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -134,14 +134,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_create_deidentify_template_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.create_deidentify_template()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{"return_value.create_dlp_job.return_value": API_RESPONSE},  # type: ignore
     )
     def test_create_dlp_job(self, get_conn):
@@ -161,14 +161,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_create_dlp_job_without_project_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.create_dlp_job()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_create_dlp_job_with_wait_until_finished(self, get_conn):
         job_for_create = DlpJob(name=DLP_JOB_PATH, state=DlpJob.JobState.PENDING)
@@ -183,7 +183,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.create_inspect_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -202,7 +202,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.create_inspect_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -221,14 +221,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_create_inspect_template_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.create_inspect_template()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.create_job_trigger.return_value": API_RESPONSE
         },  # type: ignore
@@ -247,14 +247,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_create_job_trigger_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.create_job_trigger()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.create_stored_info_type.return_value": API_RESPONSE
         },  # type: ignore
@@ -273,7 +273,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.create_stored_info_type.return_value": API_RESPONSE
         },  # type: ignore
@@ -292,14 +292,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_create_stored_info_type_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.create_stored_info_type()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.deidentify_content.return_value": API_RESPONSE
         },  # type: ignore
@@ -321,14 +321,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_deidentify_content_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.deidentify_content()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_deidentify_template_with_org_id(self, get_conn):
         self.hook.delete_deidentify_template(
@@ -343,7 +343,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_deidentify_template_with_project_id(self, get_conn):
         self.hook.delete_deidentify_template(
@@ -358,21 +358,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_deidentify_template_without_template_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_deidentify_template(template_id=None)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_deidentify_template_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_deidentify_template(template_id=TEMPLATE_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_dlp_job(self, get_conn):
         self.hook.delete_dlp_job(dlp_job_id=DLP_JOB_ID, project_id=PROJECT_ID)
@@ -382,21 +382,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_dlp_job_without_dlp_job_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_dlp_job(dlp_job_id=None, project_id=PROJECT_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_dlp_job_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_dlp_job(dlp_job_id=DLP_JOB_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_inspect_template_with_org_id(self, get_conn):
         self.hook.delete_inspect_template(
@@ -411,7 +411,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_inspect_template_with_project_id(self, get_conn):
         self.hook.delete_inspect_template(
@@ -426,21 +426,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_inspect_template_without_template_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_inspect_template(template_id=None)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_inspect_template_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_inspect_template(template_id=TEMPLATE_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_job_trigger(self, get_conn):
         self.hook.delete_job_trigger(job_trigger_id=TRIGGER_ID, project_id=PROJECT_ID)
@@ -450,21 +450,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_job_trigger_without_trigger_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_job_trigger(job_trigger_id=None, project_id=PROJECT_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_job_trigger_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_job_trigger(job_trigger_id=TRIGGER_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_stored_info_type_with_org_id(self, get_conn):
         self.hook.delete_stored_info_type(
@@ -479,7 +479,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_stored_info_type_with_project_id(self, get_conn):
         self.hook.delete_stored_info_type(
@@ -494,21 +494,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_stored_info_type_without_stored_info_type_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_stored_info_type(stored_info_type_id=None)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_delete_stored_info_type_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.delete_stored_info_type(stored_info_type_id=STORED_INFO_TYPE_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.get_deidentify_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -527,7 +527,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.get_deidentify_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -546,21 +546,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_deidentify_template_without_template_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_deidentify_template(template_id=None)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_deidentify_template_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_deidentify_template(template_id=TEMPLATE_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{"return_value.get_dlp_job.return_value": API_RESPONSE},  # type: ignore
     )
     def test_get_dlp_job(self, get_conn):
@@ -572,21 +572,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_dlp_job_without_dlp_job_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_dlp_job(dlp_job_id=None, project_id=PROJECT_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_dlp_job_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_dlp_job(dlp_job_id=DLP_JOB_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.get_inspect_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -605,7 +605,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.get_inspect_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -624,21 +624,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_inspect_template_without_template_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_inspect_template(template_id=None)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_inspect_template_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_inspect_template(template_id=TEMPLATE_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{"return_value.get_job_trigger.return_value": API_RESPONSE},  # type: ignore
     )
     def test_get_job_trigger(self, get_conn):
@@ -652,21 +652,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_job_trigger_without_trigger_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_job_trigger(job_trigger_id=None, project_id=PROJECT_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_job_trigger_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_job_trigger(job_trigger_id=TRIGGER_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.get_stored_info_type.return_value": API_RESPONSE
         },  # type: ignore
@@ -685,7 +685,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.get_stored_info_type.return_value": API_RESPONSE
         },  # type: ignore
@@ -704,21 +704,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_stored_info_type_without_stored_info_type_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_stored_info_type(stored_info_type_id=None)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_get_stored_info_type_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.get_stored_info_type(stored_info_type_id=STORED_INFO_TYPE_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{"return_value.inspect_content.return_value": API_RESPONSE},  # type: ignore
     )
     def test_inspect_content(self, get_conn):
@@ -736,14 +736,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_inspect_content_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.inspect_content()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_deidentify_templates_with_org_id(self, get_conn):
         result = self.hook.list_deidentify_templates(organization_id=ORGANIZATION_ID)
@@ -759,7 +759,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_deidentify_templates_with_project_id(self, get_conn):
         result = self.hook.list_deidentify_templates(project_id=PROJECT_ID)
@@ -775,14 +775,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_deidentify_templates_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.list_deidentify_templates()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_dlp_jobs(self, get_conn):
         result = self.hook.list_dlp_jobs(project_id=PROJECT_ID)
@@ -800,14 +800,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_dlp_jobs_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.list_dlp_jobs()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{"return_value.list_info_types.return_value": API_RESPONSE},  # type: ignore
     )
     def test_list_info_types(self, get_conn):
@@ -819,7 +819,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_inspect_templates_with_org_id(self, get_conn):
         result = self.hook.list_inspect_templates(organization_id=ORGANIZATION_ID)
@@ -835,7 +835,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_inspect_templates_with_project_id(self, get_conn):
         result = self.hook.list_inspect_templates(project_id=PROJECT_ID)
@@ -851,14 +851,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_inspect_templates_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.list_inspect_templates()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_job_triggers(self, get_conn):
         result = self.hook.list_job_triggers(project_id=PROJECT_ID)
@@ -875,14 +875,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_job_triggers_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.list_job_triggers()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_stored_info_types_with_org_id(self, get_conn):
         result = self.hook.list_stored_info_types(organization_id=ORGANIZATION_ID)
@@ -898,7 +898,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_stored_info_types_with_project_id(self, get_conn):
         result = self.hook.list_stored_info_types(project_id=PROJECT_ID)
@@ -914,14 +914,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_list_stored_info_types_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.list_stored_info_types()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{"return_value.redact_image.return_value": API_RESPONSE},  # type: ignore
     )
     def test_redact_image(self, get_conn):
@@ -940,14 +940,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_redact_image_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.redact_image()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.reidentify_content.return_value": API_RESPONSE
         },  # type: ignore
@@ -969,14 +969,14 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_reidentify_content_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.reidentify_content()
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.update_deidentify_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -997,7 +997,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.update_deidentify_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -1018,7 +1018,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_update_deidentify_template_without_template_id(self, _):
         with self.assertRaises(AirflowException):
@@ -1027,14 +1027,14 @@ class TestCloudDLPHook(unittest.TestCase):
             )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_update_deidentify_template_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.update_deidentify_template(template_id=TEMPLATE_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.update_inspect_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -1055,7 +1055,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.update_inspect_template.return_value": API_RESPONSE
         },  # type: ignore
@@ -1076,7 +1076,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_update_inspect_template_without_template_id(self, _):
         with self.assertRaises(AirflowException):
@@ -1085,14 +1085,14 @@ class TestCloudDLPHook(unittest.TestCase):
             )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_update_inspect_template_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.update_inspect_template(template_id=TEMPLATE_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.update_job_trigger.return_value": API_RESPONSE
         },  # type: ignore
@@ -1113,21 +1113,21 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_update_job_trigger_without_job_trigger_id(self, _):
         with self.assertRaises(AirflowException):
             self.hook.update_job_trigger(job_trigger_id=None, project_id=PROJECT_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_update_job_trigger_without_parent(self, _):
         with self.assertRaises(AirflowException):
             self.hook.update_job_trigger(job_trigger_id=TRIGGER_ID)
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.update_stored_info_type.return_value": API_RESPONSE
         },  # type: ignore
@@ -1148,7 +1148,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn",
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn",
         **{
             "return_value.update_stored_info_type.return_value": API_RESPONSE
         },  # type: ignore
@@ -1169,7 +1169,7 @@ class TestCloudDLPHook(unittest.TestCase):
         )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_update_stored_info_type_without_stored_info_type_id(self, _):
         with self.assertRaises(AirflowException):
@@ -1178,7 +1178,7 @@ class TestCloudDLPHook(unittest.TestCase):
             )
 
     @mock.patch(  # type: ignore
-        "airflow.contrib.hooks.gcp_dlp_hook.CloudDLPHook.get_conn"
+        "airflow.gcp.hooks.dlp.CloudDLPHook.get_conn"
     )
     def test_update_stored_info_type_without_parent(self, _):
         with self.assertRaises(AirflowException):
